1.0.1-rc2 overlaid over 1.0.0

+++ src/misc/file.c
@@ -29,6 +29,9 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #ifdef  HAVE_UNISTD_H
 #include <unistd.h>
 #endif
+#ifdef  HAVE_DIRENT_H
+#include <dirent.h>
+#endif
 #include <stdio.h>
 #include <string.h>
 #include <errno.h>
@@ -57,15 +60,6 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #elif   defined _WIN32
 #include <windows.h>                            // Sleep(), milliseconds
 #include "misc/win32.h"
-#endif
-#ifdef  HAVE_DIRENT_H
-#include <dirent.h>
-#endif
-#ifdef  HAVE_UNISTD_H
-#include <unistd.h>
-#endif
-#ifdef  _WIN32
-#include <windows.h>
 #ifndef __MINGW32__
 #include <io.h>
 #define S_ISDIR(mode) ((mode) & _S_IFDIR ? 1 : 0)
@@ -74,13 +68,7 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #endif
 #include "file.h"
 #include "misc.h"                               // getenv2()
-#include "getopt.h"                             // struct option
 
-#ifdef  DJGPP
-#ifdef  DLL
-#include "dxedll_priv.h"
-#endif
-#endif
 
 #ifndef MAXBUFSIZE
 #define MAXBUFSIZE 32768
@@ -103,6 +91,45 @@ int
 tofname (int c)
 {
   return isfname (c) ? c : '_';
+}
+
+
+const char *
+get_suffix (const char *filename)
+// Note that get_suffix() never returns NULL. Other code relies on that!
+{
+  const char *p, *s;
+
+  if (!(p = basename2 (filename)))
+    p = filename;
+  if (!(s = strrchr (p, '.')))
+    s = strchr (p, 0);                          // strchr(p, 0) and NOT "" is the
+  if (s == p)                                   //  suffix of a file without suffix
+    s = strchr (p, 0);                          // files can start with '.'
+
+  return s;
+}
+
+
+char *
+set_suffix (char *filename, const char *suffix)
+{
+  // always use set_suffix() and NEVER the code below
+  strcpy ((char *) get_suffix (filename), suffix);
+
+  return filename;
+}
+
+
+int
+filesize (const char *filename)
+{
+  struct stat fstate;
+
+  if (!stat (filename, &fstate))
+    return fstate.st_size;
+
+  return -1;
 }
 
 
@@ -154,19 +181,21 @@ realpath (const char *path, char *full_path)
 #endif
   if (*path != FILE_SEPARATOR)
     {
-      getcwd (new_path, FILENAME_MAX - 1);
+      if (getcwd (new_path, FILENAME_MAX - 1))
+        {
 #ifdef  DJGPP
-      // DJGPP's getcwd() returns a path with forward slashes
-      {
-        int l = strlen (new_path);
-        for (n = 0; n < l; n++)
-          if (new_path[n] == '/')
-            new_path[n] = FILE_SEPARATOR;
-      }
+          // DJGPP's getcwd() returns a path with forward slashes
+          {
+            int l = strlen (new_path);
+            for (n = 0; n < l; n++)
+              if (new_path[n] == '/')
+                new_path[n] = FILE_SEPARATOR;
+          }
 #endif
-      new_path += strlen (new_path);
-      if (*(new_path - 1) != FILE_SEPARATOR)
-        *new_path++ = FILE_SEPARATOR;
+          new_path += strlen (new_path);
+          if (*(new_path - 1) != FILE_SEPARATOR)
+            *new_path++ = FILE_SEPARATOR;
+        }
     }
   else
     {
@@ -418,7 +447,7 @@ basename2 (const char *path)
   if (p2 > p1)                                  // use the last separator in path
     p1 = p2;
 #else
-  p1 = strrchr (path, FILE_SEPARATOR);
+  p1 = strrchr ((char *) path, FILE_SEPARATOR);
 #endif
 
 #if     defined DJGPP || defined __CYGWIN__ || defined _WIN32
@@ -427,33 +456,6 @@ basename2 (const char *path)
 #endif
 
   return p1 ? p1 + 1 : path;
-}
-
-
-const char *
-get_suffix (const char *filename)
-// Note that get_suffix() never returns NULL. Other code relies on that!
-{
-  const char *p, *s;
-
-  if (!(p = basename2 (filename)))
-    p = filename;
-  if (!(s = strrchr (p, '.')))
-    s = strchr (p, 0);                          // strchr(p, 0) and NOT "" is the
-  if (s == p)                                   //  suffix of a file without suffix
-    s = strchr (p, 0);                          // files can start with '.'
-
-  return s;
-}
-
-
-char *
-set_suffix (char *filename, const char *suffix)
-{
-  // always use set_suffix() and NEVER the code below
-  strcpy ((char *) get_suffix (filename), suffix);
-
-  return filename;
 }
 
 
@@ -507,7 +509,7 @@ same_file (const char *filename1, const char *filename2)
 
 
 int
-same_filesystem (const char *filename1, const char *filename2)
+same_fs (const char *filename1, const char *filename2)
 // returns 1 if filename1 and filename2 reside on one file system, 0 if not
 //  (or an error occurred)
 {
@@ -588,7 +590,7 @@ rename2 (const char *oldname, const char *newname)
   dirname2 (newname, dir2);
 
   // We should use dirname{2}() in case oldname or newname doesn't exist yet
-  if (same_filesystem (dir1, dir2))
+  if (same_fs (dir1, dir2))
     {
       if (access (newname, F_OK) == 0 && !same_file (oldname, newname))
         {
@@ -600,7 +602,7 @@ rename2 (const char *oldname, const char *newname)
     }
   else
     {
-      retval = fcopy (oldname, 0, fsizeof (oldname), newname, "wb");
+      retval = fcopy (oldname, 0, filesize (oldname), newname, "wb");
       // don't remove unless the file can be copied
       if (retval == 0)
         {
@@ -617,7 +619,7 @@ rename2 (const char *oldname, const char *newname)
 int
 truncate2 (const char *filename, unsigned long new_size)
 {
-  unsigned long size = fsizeof (filename);
+  unsigned long size = filesize (filename);
   struct stat fstate;
 
   stat (filename, &fstate);
@@ -643,41 +645,11 @@ truncate2 (const char *filename, unsigned long new_size)
         }
 
       fclose (file);
-    }
-  else
-    truncate (filename, new_size);
 
-  return 0;                                     // success
-}
-
-
-int
-fsizeof (const char *filename)
-{
-  struct stat fstate;
-
-  if (!stat (filename, &fstate))
-    return fstate.st_size;
-
-  return -1;
-}
-
-
-char *
-baknam (char *fname)
-{
-  char suffix[8];
-  int i = 1;
-
-  set_suffix (fname, ".bak");
-  while (!access (fname, F_OK))
-    {
-      sprintf (suffix, ".b%02d", i);
-      set_suffix (fname, suffix);
-      i++;
+      return 0; // success
     }
 
-  return fname;
+  return truncate (filename, new_size);
 }
 
 
@@ -736,258 +708,105 @@ fcopy (const char *source, size_t start, size_t len, const char *dest, const cha
 }
 
 
-int
-quick_io_c (int value, size_t pos, const char *filename, const char *mode)
+unsigned char *
+file_get_contents (const char *filename, int maxlength)
 {
-  int result;
-  FILE *fh;
+  FILE *fh = NULL;
+  unsigned char *p = NULL;
+  int len = filesize (filename);        
 
-  if (!(fh = fopen (filename, (const char *) mode)))
-    return -1;
+  if (len < 0)
+    return NULL;
 
-  fseek (fh, pos, SEEK_SET);
+  if (len == 0)
+    return (unsigned char *) "";
 
-  if (*mode == 'r' && mode[1] != '+')           // "r+b" always writes
-    result = fgetc (fh);
-  else
-    result = fputc (value, fh);
+  if (len > maxlength)
+    len = maxlength;  
+
+  if (!(fh = fopen (filename, "rb")))
+    return NULL;
+
+  if (!(p = (unsigned char *) malloc (len + 2)))
+    {
+      fclose (fh);
+      return NULL;
+    } 
+
+  len = fread (p, 1, len, fh);
+
+  if (!len)
+    {
+      free (p);
+      p = NULL;
+    }
+
+  // terminate anyways
+  p[len] = 0;
 
   fclose (fh);
 
-  return result;
+  return p;
 }
 
 
 int
-quick_io (void *buffer, size_t start, size_t len, const char *filename,
-          const char *mode)
+mkdir2_func (const char *path, int mode)
 {
-  int result;
-  FILE *fh;
+  int result = 0;
+  char dir[FILENAME_MAX], *p = NULL;
 
-  if (!(fh = fopen (filename, (const char *) mode)))
-    return -1;
-
-  fseek (fh, start, SEEK_SET);
-
-  // Note the order of arguments of fread() and fwrite(). Now quick_io()
-  //  returns the number of characters read or written. Some code relies on
-  //  this behaviour!
-  if (*mode == 'r' && mode[1] != '+')           // "r+b" always writes
-    result = (int) fread (buffer, 1, len, fh);
-  else
-    result = (int) fwrite (buffer, 1, len, fh);
-
-  fclose (fh);
-  return result;
-}
-
-
-static int
-getfile_recursion (const char *fname, int (*callback_func) (const char *),
-                        int *calls, int flags)
-{
-  char path[FILENAME_MAX];
-  struct stat fstate;
-
-  if (strlen (fname) >= FILENAME_MAX - 2)
-    return 0;
-
-  realpath2 (fname, path);
-
-  /*
-    Try to get file status information only if the file with name fname exists.
-    If the file does not exist I set st_mode to 0 instead of __S_IFREG, because I
-    don't know if the latter is portable. - dbjh
-  */
-  if (access (path, F_OK) == 0)
-    {
-      if (stat (path, &fstate) != 0)
-        return 0;
-    }
-  else
-    fstate.st_mode = 0;
-
-  /*
-    We test whether fname is a directory, because we handle directories
-    differently. The callback function should test whether its argument is a
-    regular file, a character special file, a block special file, a FIFO
-    special file, a symbolic link or a socket. If the flags
-    GETFILE_FILES_ONLY, GETFILE_RECURSIVE and
-    GETFILE_RECURSIVE_ONCE were not used by the calling function, it may
-    also have to test whether the argument is a directory.
-  */
-  if (S_ISDIR (fstate.st_mode) ?
-        !(flags & (GETFILE_FILES_ONLY |
-                   GETFILE_RECURSIVE |
-                   GETFILE_RECURSIVE_ONCE)) :
-        1)                                      // everything else: call callback
-    {
-      int result = 0; 
-
-#ifdef  DEBUG
-      printf ("callback_func() == %s\n", path);
-      fflush (stdout);
-#endif
-
-      result = callback_func (path);
-
-      if (!result)
-        (*calls)++;
-
-      return result;
-    }
-
-  if (S_ISDIR (fstate.st_mode) &&
-      (flags & (GETFILE_RECURSIVE | GETFILE_RECURSIVE_ONCE)))
-    {
-      int result = 0; 
-      struct dirent *ep;
-      DIR *dp;
-      char buf[FILENAME_MAX], *p;
-
-#if     defined __MSDOS__ || defined _WIN32 || defined __CYGWIN__
-      char c = toupper (path[0]);
-      if (path[strlen (path) - 1] == FILE_SEPARATOR ||
-          (c >= 'A' && c <= 'Z' && path[1] == ':' && path[2] == 0))
-#else
-      if (path[strlen (path) - 1] == FILE_SEPARATOR)
-#endif
-        p = "";
-      else
-        p = FILE_SEPARATOR_S;
-
-      if ((dp = opendir (path)))
-        {
-          while ((ep = readdir (dp)))
-            if (strcmp (ep->d_name, ".") != 0 &&
-                strcmp (ep->d_name, "..") != 0)
-              {
-                sprintf (buf, "%s%s%s", path, p, ep->d_name);
-                result = getfile_recursion (buf, callback_func, calls,
-                           flags & ~GETFILE_RECURSIVE_ONCE);
-                if (result != 0)
-                  break;
-              }
-          closedir (dp);
-        }
-    }
-
-  return 0;
-}
-
-
-int
-getfile (int argc, char **argv, int (*callback_func) (const char *), int flags)
-{
-  int x = optind, calls = 0, result = 0;
-
-  for (; x < argc; x++)
-    {
-      result = getfile_recursion (argv[x], callback_func, &calls, flags);
-      if (result != 0)
-        break;
-    }
-
-  return calls;
-}
-
-
-int
-mkdir2 (const char *name)
-// create a directory and check its permissions
-{
-  struct stat *st = NULL;
-
-  if (stat (name, st) == -1)
-    {
-      if (errno != ENOENT)
-        {
-          fprintf (stderr, "stat %s", name);
-          return -1;
-        }
-      if (mkdir (name, 0700) == -1)
-        {
-          fprintf (stderr, "mkdir %s", name);
-          return -1;
-        }
-      if (stat (name, st) == -1)
-        {
-          fprintf (stderr, "stat %s", name);
-          return -1;
-        }
-    }
-
-  if (!S_ISDIR (st->st_mode))
-    {
-      fprintf (stderr, "%s is not a directory\n", name);
-      return -1;
-    }
-  if (st->st_uid != getuid ())
-    {
-      fprintf (stderr, "%s is not owned by you\n", name);
-      return -1;
-    }
-  if (st->st_mode & 077)
-    {
-      fprintf (stderr, "%s must not be accessible by other users\n", name);
-      return -1;
-    }
-
-  return 0;
-}
-
-
-int
-makepath_func (const char *path)
-{
-  int len;
-  char *p;
-  char dir[FILENAME_MAX];
-
-  /* Skip initial slashes */
   while (*path == '/')
     path++;
 
-  p = strchr (path, '/');
+  strncpy (dir, path, FILENAME_MAX)[FILENAME_MAX - 1] = 0;
 
-  /* Done, path is now a filename */
-  if (!p)
-    return 1;
+  if ((p = strchr (dir, '/')))
+    *p = 0;
 
-  len = p - path;
-  if (len >= FILENAME_MAX)
-    {
-      fprintf (stderr, "ERROR: directory name too long\n");
-      return 0;
-    }
+  result = mkdir (dir, mode);
+// returns also an error if dir exists already
+//  if (result == -1)
+//    return -1;
 
-  strncpy (dir, path, len);
-  dir[len] = '\0';
+  if (chdir (dir) == -1)
+    return -1;
 
-  if (mkdir2 (dir) == -1)
+  if (!(p = strchr ((char *) path, '/')))
     return 0;
-  chdir (dir);
-
-  return makepath (p + 1);
+  return mkdir2_func (p, mode);
 }
 
 
 int
-makepath (const char *path)
+mkdir2 (const char *path, int mode)
 {
-  char savedcwd[FILENAME_MAX];
-  int ret;
+  // TODO: make mkdir2() work the same way as rmdir2() instead of recursion
+  char buf[FILENAME_MAX];
+  int result = 0;
 
-  getcwd (savedcwd, FILENAME_MAX);
-  ret = makepath_func (path);
-  chdir (savedcwd);
+  if (strlen (path) >= FILENAME_MAX)   
+    {
+      fprintf (stderr, "ERROR: directory name too long\n");
+      return -1;
+    }
 
-  return ret;
+  if (!getcwd (buf, FILENAME_MAX))
+    return -1;
+    
+  if (*path == '/')
+    if (chdir ("/") == -1)
+      return -1;
+
+  result = mkdir2_func (path, mode);
+
+  if (chdir (buf) == -1)
+    return -1;
+
+  return result;
 }
 
 
-#if 0
 int
 rmdir2 (const char *path)
 {
@@ -998,9 +817,12 @@ rmdir2 (const char *path)
 
   if (!(dp = opendir (path)))
     return -1;
+ 
+  if (!getcwd (cwd, FILENAME_MAX))
+    return -1;
 
-  getcwd (cwd, FILENAME_MAX);
-  chdir (path);
+  if (chdir (path) == -1)
+    return -1;
 
   while ((ep = readdir (dp)) != NULL)
     {
@@ -1018,35 +840,21 @@ rmdir2 (const char *path)
     }
 
   closedir (dp);
-  chdir (cwd);
 
-  return rmdir (path);
+  if (chdir (cwd) == -1)
+    return -1;
+ 
+  return rmdir (path);   
+}
+
+
+//#if 0
+#ifdef  TEST
+int
+main (int argc, char **argv)
+{
+  mkdir2 ("/test/test2", 0755);
+
+  return 0;
 }
 #endif
-
-
-unsigned char *
-fread2 (const char *filename, int maxlength)
-{
-  FILE *fh = NULL;
-  unsigned char *p = NULL;
-  int len = fsizeof (filename);
-
-  if (len > maxlength || len == -1)
-    return NULL;
-
-  if (!(fh = fopen (filename, "rb")))
-    return NULL;
-
-  if (!(p = (unsigned char *) malloc (len + 1)))
-    {
-      fclose (fh);
-      return NULL;
-    }
-
-  fread (p, len, 1, fh);
-
-  fclose (fh);
-
-  return p;
-}
